+++ src/main/distance.ino
@@ -2,7 +2,7 @@
 
 // 2点の緯度/経度から距離を計算
 // https://keisan.casio.jp/exec/system/1257670779
-double distance(float lng1, float lat1, float lng2, float lat2){// x1, y1, x2, y2
+double distance(double lng1, double lat1, double lng2, double lat2){// x1, y1, x2, y2
   double x1, y1, x2, y2;
   double d;
   double sin_y1, sin_y2, cos_y1, cos_y2, cos_dx, arc_cos;
